extract consensus wrapping code to processor
